+++ tests/LaravelEnvcoderTest.php
@@ -17,6 +17,12 @@ class LaravelEnvcoderTest extends Orchestra\Testbench\TestCase
         return ['harmonic\LaravelEnvcoder\LaravelEnvcoderServiceProvider'];
     }
 
+    public function setUp() : void
+    {
+        parent::setUp();
+        Config::set('envcoder.resolve', 'merge');
+    }
+
     /**
      * Resolve application Console Kernel implementation.
      *
@@ -120,6 +126,35 @@ class LaravelEnvcoderTest extends Orchestra\Testbench\TestCase
     }
 
     /**
+     * Can encypt a .env.testing file.
+     *
+     * @test
+     * @return void
+     */
+    public function canEncryptEnvTesting()
+    {
+        // Arrange
+        $this->createEnvFileWithPassword();
+        copy('.env', '.env.testing');
+        unlink('.env');
+
+        // Act
+        $this->artisan('env:encrypt -p password --source .env.testing');
+        copy('.env.testing', '.env.testing.original');
+        unlink('.env.testing');
+        $this->artisan('env:decrypt -p password --source .env.testing.enc');
+
+        // Assert
+        $this->assertTrue(file_exists('.env.testing.enc'));
+        $this->assertTrue(file_exists('.env.testing'));
+        $this->assertEquals(file_get_contents('.env.testing.original'), file_get_contents('.env.testing'));
+
+        unlink('.env.testing');
+        unlink('.env.testing.enc');
+        unlink('.env.testing.original');
+    }
+
+    /**
      * Can use password option and prompt for password.
      *
      * @test
@@ -164,7 +199,7 @@ class LaravelEnvcoderTest extends Orchestra\Testbench\TestCase
         // Arrange
         $this->createEnvFile();
         File::encryptFileWithPassword('.env', '.env.enc', 'password');
-        Config::set('laravel-envcoder.resolve', 'overwrite');
+        Config::set('envcoder.resolve', 'overwrite');
         $originalEnv = file_get_contents('.env');
 
         // Act
@@ -184,7 +219,7 @@ class LaravelEnvcoderTest extends Orchestra\Testbench\TestCase
     public function canDecryptIgnore()
     {
         // Arrange
-        Config::set('laravel-envcoder.resolve', 'ignore');
+        Config::set('envcoder.resolve', 'ignore');
         $this->createEnvFile();
         $originalEnv = 'VAR1=TEST'.PHP_EOL.'VAR2=TEST2'.PHP_EOL;
 
@@ -212,7 +247,7 @@ class LaravelEnvcoderTest extends Orchestra\Testbench\TestCase
     {
         // Arrange
         $this->createEnvFile();
-        Config::set('laravel-envcoder.resolve', 'merge');
+        Config::set('envcoder.resolve', 'merge');
 
         $env2 = fopen('.env2', 'w');
         fwrite($env2, 'VAR3=TEST3'.PHP_EOL.'VAR4=TEST4'.PHP_EOL);
@@ -242,7 +277,7 @@ class LaravelEnvcoderTest extends Orchestra\Testbench\TestCase
     {
         // Arrange
         $this->createEnvFileWithPassword();
-        Config::set('laravel-envcoder.resolve', 'prompt');
+        Config::set('envcoder.resolve', 'prompt');
 
         $env2 = fopen('.env2', 'w');
         fwrite($env2, 'VAR1=TEST3'.PHP_EOL.'VAR3=TEST4'.PHP_EOL.'ENV_PASSWORD=password');